+++ packages/xchain-bitcoin/__mocks__/sochain.ts
@@ -33,8 +33,25 @@ export default {
 
     //Mock get_unspent_txs
     mock.onGet(/\/get_tx_unspent\//).reply(function (config: MockConfig) {
+      const split = config.url?.split('/')
+
+      //the address is always the 7th, the optional 8th param would bethe startgin from txid to allow paging
+      const address = split?.[7] || ''
+      const startingfromTxId = split?.length == 9 ? split?.[8] : ''
+
+      let filePath = `./response/unspent-txs/${address}.json`
+      if (startingfromTxId) {
+        // this allows you to page utxos startign from a given txid
+        filePath = `./response/unspent-txs/${address}/${startingfromTxId}.json`
+      }
+      const resp = require(filePath)
+      return [200, resp]
+    })
+
+    //Mock is_tx_confirmed
+    mock.onGet(/\/is_tx_confirmed\//).reply(function (config: MockConfig) {
       const id = config.url?.split('/').pop() ?? ''
-      const resp = require(`./response/unspent-txs/${id}.json`)
+      const resp = require(`./response/is-tx-confirmed/${id}.json`)
       return [200, resp]
     })