+++ packages/xchain-litecoin/__tests__/utils.test.ts
@@ -14,6 +14,7 @@ describe('Litecoin Utils Test', () => {
   const utxo = {
     hash: '7fc1d2c1e4017a6aea030be1d4f5365d11abfd295f56c13615e49641c55c54b8',
     index: 0,
+    value: witness.value,
     witnessUtxo: witness,
     txHex:
       '01000000000101233b5e27c30135274523c69c68558dddd265e63d9f2db1953e59c6ba6dc4912e0100000000ffffffff01dc410f0000000000160014ea0b3a147753eaf29d8aa820b335876daa0d61cb02483045022100c324931915f3215cbc4175e196a78b11333dcb08bc929c417bc98645acd638fc022028bb7bbb5da72f630aeba29a76a763407c3a98a7e8809c78ffab02f2d2a4eb0e012102dbc2fa9261379482e9ed484dc2c8b8a3ca7543391de90159a51e1791c4d2271b00000000',
@@ -43,5 +44,16 @@ describe('Litecoin Utils Test', () => {
     expect(estimates.fast).toBeDefined()
     expect(estimates.fastest).toBeDefined()
     expect(estimates.average).toBeDefined()
+  })
+  it('should fetch as many uxtos as are associated with an address', async () => {
+    const address = 'M8T1B2Z97gVdvmfkQcAtYbEepune1tzGua'
+    const utxos: UTXOs = await Utils.scanUTXOs({
+      sochainUrl: 'https://sochain.com/api/v2',
+      network: 'mainnet',
+      address,
+    })
+    expect(utxos.length).toEqual(213)
+    expect(utxos?.[0].hash).toEqual('65b34acff41570854758adf6bdafc94c0c33f78194d7f49f1cf8d24314b4d47a')
+    expect(utxos?.[212].hash).toEqual('f180c1cd0a8e719456f3f033c497bae2cedc482d87443b668c0a5a277272b2ba')
   })
 })